+++ constants.php
@@ -1,6 +1,15 @@
 <?php
 /**
  * PHP_BS_GRID constants
+ *
+ * php datagrid with jquery, jquery-ui and bootstrap frontend
+ * https://github.com/pontikis/php_bs_grid
+ *
+ * @author     Christos Pontikis http://pontikis.net
+ * @copyright  Christos Pontikis
+ * @license    MIT http://opensource.org/licenses/MIT
+ * @version    0.9.5 (XX Jul 2017)
+ *
  */
 
 // pagination
@@ -26,12 +35,6 @@ define('C_PHP_BS_GRID_CRITERIA_TEXT_STARTS_WITH', 3);
 define('C_PHP_BS_GRID_CRITERIA_TEXT_CONTAINS', 4);
 define('C_PHP_BS_GRID_CRITERIA_TEXT_IS_NULL', 5);
 
-// lookup criteria operators
-define('C_PHP_BS_GRID_CRITERIA_LOOKUP_IGNORE', 1);
-define('C_PHP_BS_GRID_CRITERIA_LOOKUP_EQUAL', 2);
-define('C_PHP_BS_GRID_CRITERIA_LOOKUP_NOT_EQUAL', 3);
-define('C_PHP_BS_GRID_CRITERIA_LOOKUP_IS_NULL', 4);
-
 // number criteria operators
 define('C_PHP_BS_GRID_CRITERIA_NUMBER_IGNORE', 1);
 define('C_PHP_BS_GRID_CRITERIA_NUMBER_EQUAL', 2);
@@ -49,6 +52,19 @@ define('C_PHP_BS_GRID_CRITERIA_DATE_LESS_THAN_OR_EQUAL_TO', 4);
 define('C_PHP_BS_GRID_CRITERIA_DATE_GREATER_THAN', 5);
 define('C_PHP_BS_GRID_CRITERIA_DATE_GREATER_THAN_OR_EQUAL_TO', 6);
 define('C_PHP_BS_GRID_CRITERIA_DATE_IS_NULL', 7);
+
+// lookup criteria operators
+define('C_PHP_BS_GRID_CRITERIA_LOOKUP_IGNORE', 1);
+define('C_PHP_BS_GRID_CRITERIA_LOOKUP_EQUAL', 2);
+define('C_PHP_BS_GRID_CRITERIA_LOOKUP_NOT_EQUAL', 3);
+define('C_PHP_BS_GRID_CRITERIA_LOOKUP_IS_NULL', 4);
+
+// autocomplete criteria operators
+define('C_PHP_BS_GRID_CRITERIA_AUTOCOMPLETE_EQUAL', 1);
+define('C_PHP_BS_GRID_CRITERIA_AUTOCOMPLETE_IS_NULL', 2);
+
+define('C_PHP_BS_GRID_CRITERIA_AUTOCOMPLETE_DISPLAY_IS_NULL_YES', 1);
+define('C_PHP_BS_GRID_CRITERIA_AUTOCOMPLETE_DISPLAY_IS_NULL_NO', 2);
 
 // misc
 define('C_PHP_BS_GRID_DACAPO_SQL_PLACEHOLDER', '?');